Implement naive PrivateSettingsContext
